Move integration tests to integration folder
- Moved re-usable fixtures to conftest.py
- Renamed test_integration.py to test_query.py
